+++ docs/guides/changelog.md
@@ -9,6 +9,7 @@ description: |-
 
 ## Unreleased
 
+- Add `nxos_snmp` resource and data source
 - Add UDLD configuration to `nxos_system` resource and data source
 - Add NTP source interface configuration to `nxos_ntp` resource and data source
 - Add NTP access group configuration to `nxos_ntp` resource and data source